+++ src/renderables/WorldObject.h
@@ -13,7 +13,7 @@ class WorldObject:public Object{
   float vertices[6];
   WorldObject (float xSize, float ySize, float zSize, string id);
   void generateVertices(float *buffer);
-  void setPos(float xPos, float yPos, float zPos);
+  void setPos(int xPos, int yPos, int zPos);
   void translate(float xDistance, float yDistance, float zDistance);
   void setTransform(float xTransform, float yTransform, float zTransform);
   void transform(float xFactor, float yFactor, float zFactor);
